Re-baseline tests to run successfully when .NET 4.5 is installed
.NET 4.5 changed the way sorting works. Assuming all items have an
absolute order, this behavior is transparent for consumers of
Array.Sort. However, for cases where a series of elements are considered
equivalent, the new implementation returns a different, albeit, valid
sort order for the given input.

This breaks our test cases which compare the execution plan and query
results against a checked-in baseline. This change updates the baseline
so that developers using Visual Studio 2012 can run tests successfully.

Please note that although NQuery still targets 2.0 the unit test runner
will (unfortunately) run them against 4.5. Needless to say, running the
test using Visual Studio 2010 installed on a machine which only has .NET
4.0, those test will now fail. Hence, I've marked the test project as
targeting .NET 4.5 so that this doesn't happen accidentally.

+++ Src/NQuery.Tests/Tests/AggregateTests.cs
@@ -117,16 +117,16 @@ namespace NQuery.Tests
 			DataTable dataTable = query.ExecuteDataTable();
 
 			Assert.AreEqual(typeof(int), dataTable.Columns["Int32Sum"].DataType);
-			Assert.AreEqual(typeof(int), dataTable.Columns["Int16Sum"].DataType);
+			Assert.AreEqual(typeof(long), dataTable.Columns["Int16Sum"].DataType);
 			Assert.AreEqual(typeof(long), dataTable.Columns["Int64Sum"].DataType);
 			
 			Assert.AreEqual(2, dataTable.Rows.Count);
 			Assert.AreEqual(1, dataTable.Rows[0]["Int32Sum"]);
-			Assert.AreEqual(4, dataTable.Rows[0]["Int16Sum"]);
+			Assert.AreEqual((long)4, dataTable.Rows[0]["Int16Sum"]);
 			Assert.AreEqual((long)4, dataTable.Rows[0]["Int64Sum"]);
 
 			Assert.AreEqual(4, dataTable.Rows[1]["Int32Sum"]);
-			Assert.AreEqual(11, dataTable.Rows[1]["Int16Sum"]);
+			Assert.AreEqual((long)11, dataTable.Rows[1]["Int16Sum"]);
 			Assert.AreEqual((long)11, dataTable.Rows[1]["Int64Sum"]);
 		}
 	}
